Move some classes and functions from the gui module
